refactor: メンバー一覧コンポーネントを改修
