+++ src/pages/member.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react";
 import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/dist/ScrollTrigger";
-import MemberWindow from "../ecosystem/member/memberWindow";
+import MemberWindow from "../ecosystem/member/";
 import AngeWindow from "../ecosystem/member/angeWindow";
 import LizeWindow from "../ecosystem/member/lizeWindow";
 import InuiWindow from "../ecosystem/member/inuiWindow";
@@ -16,7 +16,7 @@ const Member: React.FC = () => {
   const windows = [
     {
       key: "member",
-      content: <MemberWindow isRight={false} />,
+      content: <MemberWindow isTop={true} />,
     },
     {
       key: "ange",
