+++ src/template.scss
@@ -1,0 +1,7 @@
+$ange_font_color: rgba($color: #C83C35, $alpha: 0.8);
+$lize_font_color: rgba($color: #42FFFF, $alpha: 0.8);
+$inui_font_color: rgba($color: #9D3757, $alpha: 0.8);
+
+$ange_background_color: rgba($color: #C83C35, $alpha: 0.5);
+$lize_background_color: rgba($color: #42FFFF, $alpha: 0.5);
+$inui_background_color: rgba($color: #9D3757, $alpha: 0.5);
